requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
requests==2.31.0
pycryptodome